Move accordion keydown handler to trigger
