+++ packages/react/src/accordion/item/AccordionItem.test.tsx
@@ -1,35 +1,14 @@
 import * as React from 'react';
 import { Accordion } from '@base-ui-components/react/accordion';
 import { describeConformance, createRenderer } from '#test-utils';
-import { NOOP } from '../../utils/noop';
-import { AccordionRootContext } from '../root/AccordionRootContext';
-
-const accordionRootContextValue: AccordionRootContext = {
-  accordionItemRefs: { current: [] },
-  direction: 'ltr',
-  disabled: false,
-  handleValueChange: NOOP,
-  hiddenUntilFound: false,
-  keepMounted: false,
-  orientation: 'vertical',
-  state: {
-    value: [0],
-    disabled: false,
-    orientation: 'vertical',
-  },
-  value: [0],
-};
 
 describe('<Accordion.Item />', () => {
   const { render } = createRenderer();
 
   describeConformance(<Accordion.Item />, () => ({
-    render: (node) =>
-      render(
-        <AccordionRootContext.Provider value={accordionRootContextValue}>
-          {node}
-        </AccordionRootContext.Provider>,
-      ),
+    render: (node) => {
+      return render(<Accordion.Root>{node}</Accordion.Root>);
+    },
     refInstanceof: window.HTMLDivElement,
   }));
 });
