+++ packages/react/src/accordion/root/AccordionRootContext.ts
@@ -11,6 +11,7 @@ export interface AccordionRootContext {
   handleValueChange: (newValue: number | string, nextOpen: boolean) => void;
   hiddenUntilFound: boolean;
   keepMounted: boolean;
+  loop: boolean;
   orientation: Orientation;
   state: AccordionRoot.State;
   value: AccordionValue;
